Move Devbox Services functions to services.go
